+++ commands/setLanguage.js
@@ -1,5 +1,6 @@
 const vscode = require('vscode');
 
+// set langauge mode to SQL
 const setSQL = vscode.commands.registerCommand('edctools.setSQLLanguage', () => {
     const editor = vscode.window.activeTextEditor;
     if (editor) {
@@ -7,6 +8,7 @@ const setSQL = vscode.commands.registerCommand('edctools.setSQLLanguage', () => 
     }
 });
 
+// set language mode to Powershell
 const setPowerShell = vscode.commands.registerCommand('edctools.setPowerShellLanguage', () => {
     const editor = vscode.window.activeTextEditor;
     if (editor) {