+++ extension.js
@@ -1,14 +1,13 @@
 const { moveTabFirst, moveTabLast } = require('./commands/manipulateTabs');
 const { openPanes, closePanes } = require('./commands/manipulatePanes');
-const { setPowerShell, setSQL} = require('./commands/setLanguage')
-
+const { setPowerShell, setSQL } = require('./commands/setLanguage')
+const { enhancedQuickOpen } = require('./commands/enhancedQuickOpen')
 
 function activate(context) {
 
-    context.subscriptions.push(setSQL, setPowerShell, openPanes, closePanes, moveTabFirst, moveTabLast);
+    context.subscriptions.push(setSQL, setPowerShell, openPanes, closePanes, moveTabFirst, moveTabLast, enhancedQuickOpen);
 
 }
-
 
 function deactivate() {}