+++ CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 ## [Unreleased]
 
 - fix documentation on tab movement keystrokes
+- source code reorganization
+- added enhancedQuickOpen function
 
 ## [0.0.3] - 2025-10-18